- added support for multipart and binary data in microflow services - fixed deserialization of integers - fixed sending parameters in combination with multipart data - wrote readme section about Rest services compared to app/web services
git-svn-id: https://teamserver.sprintr.com/2f1dfdec-5e3d-497c-9706-741732a51cf2/trunk@161 5f19bcb0-64dc-40a8-b24f-70cbb7b2b6a4

+++ javasource/tests/actions/multiplefile.java
@@ -41,7 +41,7 @@ public class multiplefile extends UserAction<Boolean>
 		byte[] buffer = new byte[size];
 		IOUtils.read(Core.getFileDocumentContent(getContext(), __file), buffer);
 		
-		byte[] resultbuffer = new byte[size * file.getMultiplier()];
+		byte[] resultbuffer = new byte[size * (int)file.getMultiplier()];
 		
 		//MWE: worst implementation ever, but thats not the point :)
 		for(int i = 0; i < file.getMultiplier(); i++)
